rename webhook config directory
