+++ app/domain/housekeeper/housekeeper.go
@@ -13,7 +13,7 @@ import (
 
 	"github.com/rs/zerolog/log"
 
-	config "github.com/cloudzero/cloudzero-agent/app/config/insights-controller"
+	config "github.com/cloudzero/cloudzero-agent/app/config/webhook"
 	"github.com/cloudzero/cloudzero-agent/app/types"
 	"github.com/cloudzero/cloudzero-agent/app/utils"
 )
