+++ app/types/resource.go
@@ -6,7 +6,7 @@ package types
 import (
 	"time"
 
-	"github.com/cloudzero/cloudzero-agent/app/config/insights-controller"
+	config "github.com/cloudzero/cloudzero-agent/app/config/webhook"
 )
 
 type (
